+++ directed.go
@@ -116,6 +116,21 @@ func (d *directed[K, T]) AddEdge(sourceHash, targetHash K, options ...func(*Edge
 	return d.addEdge(sourceHash, targetHash, edge)
 }
 
+func (d *directed[K, T]) AddEdgesFrom(g Graph[K, T]) error {
+	edges, err := g.Edges()
+	if err != nil {
+		return fmt.Errorf("failed to get edges: %w", err)
+	}
+
+	for _, edge := range edges {
+		if err := d.AddEdge(copyEdge(edge)); err != nil {
+			return fmt.Errorf("failed to add (%v, %v): %w", edge.Source, edge.Target, err)
+		}
+	}
+
+	return nil
+}
+
 func (d *directed[K, T]) Edge(sourceHash, targetHash K) (Edge[T], error) {
 	edge, err := d.store.Edge(sourceHash, targetHash)
 	if err != nil {
@@ -141,6 +156,10 @@ func (d *directed[K, T]) Edge(sourceHash, targetHash K) (Edge[T], error) {
 			Data:       edge.Properties.Data,
 		},
 	}, nil
+}
+
+func (d *directed[K, T]) Edges() ([]Edge[K], error) {
+	return d.store.ListEdges()
 }
 
 func (d *directed[K, T]) RemoveEdge(source, target K) error {
@@ -250,4 +269,8 @@ func (d *directed[K, T]) edgesAreEqual(a, b Edge[T]) bool {
 	bTargetHash := d.hash(b.Target)
 
 	return aSourceHash == bSourceHash && aTargetHash == bTargetHash
+}
+
+func copyEdge[K comparable](edge Edge[K]) (K, K, func(properties *EdgeProperties)) {
+	return edge.Source, edge.Target, copyEdgeProperties(edge.Properties)
 }