+++ CHANGELOG.md
@@ -8,7 +8,9 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [0.20.0] - 2023-05-01
 
 ### Added
-* Added the `AddVerticesFrom` method for adding all vertices from another graph.
+* Added the `Graph.AddVerticesFrom` method for adding all vertices from another graph.
+* Added the `Graph.AddEdgesFrom` method for adding all edges from another graph.
+* Added the `Graph.Edges` method for obtaining all edges as a slice.
 
 ## [0.19.0] - 2023-04-23